+++ src/modules/auth/api/auth.controller.ts
@@ -1,7 +1,7 @@
 import { Body, Controller, Get, Headers, HttpCode, Ip, Post, Res, UseGuards } from '@nestjs/common';
 import { HTTP_Status } from '../../../main/enums/http-status.enum';
 import { RegisterInputDto } from './input-dto/register.input.dto';
-import { ApiExcludeEndpoint, ApiTags } from '@nestjs/swagger';
+import { ApiTags } from '@nestjs/swagger';
 import { CommandBus } from '@nestjs/cqrs';
 import { ConfirmationCodeInputDto } from './input-dto/confirmation-code.input.dto';
 import { PasswordRecoveryInputDto } from './input-dto/password-recovery.input.dto';
@@ -37,7 +37,7 @@ import {
   SwaggerDecoratorsByRegistration,
   SwaggerDecoratorsByRegistrationEmailResending,
   SwaggerDecoratorsByUpdateTokens,
-} from './swagger.auth.decorators';
+} from '../swagger/swagger.auth.decorators';
 import { UpdateTokensCommand } from '../application/use-cases/update-tokens.use-case';
 import { CurrentUserId } from '../../../main/decorators/user.decorator';
 import { JwtAuthGuard } from './guards/jwt-auth.guard';
@@ -127,7 +127,7 @@ export class AuthController {
    * @param body
    */
   @SwaggerDecoratorsByPasswordRecovery()
-  // @UseGuards(GoogleEnterpriseRecaptchaGuard)
+  @UseGuards(GoogleEnterpriseRecaptchaGuard)
   @Post('password-recovery')
   @HttpCode(HTTP_Status.NO_CONTENT_204)
   async passwordRecovery(@Body() body: PasswordRecoveryInputDto): Promise<null> {
@@ -181,6 +181,13 @@ export class AuthController {
     return notification.getData();
   }
 
+  /**
+   * @description Update tokens
+   * @param sessionData
+   * @param ip
+   * @param deviceName
+   * @param res
+   */
   @SwaggerDecoratorsByUpdateTokens()
   @Post('update-tokens')
   @UseGuards(RefreshTokenGuard)
@@ -201,6 +208,10 @@ export class AuthController {
     return { accessToken };
   }
 
+  /**
+   * @description Get user info by id
+   * @param userId
+   */
   @SwaggerDecoratorsByMe()
   @Get('me')
   @UseGuards(JwtAuthGuard)
@@ -208,17 +219,5 @@ export class AuthController {
     const user = await this.usersQueryRepository.findUserById(userId);
     if (!user) return;
     return new MeViewDto(user);
-  }
-
-  //need for testing recaptcha --- > remove in production
-  @ApiExcludeEndpoint()
-  @UseGuards(GoogleEnterpriseRecaptchaGuard)
-  @Post('password-recovery-test')
-  @HttpCode(HTTP_Status.NO_CONTENT_204)
-  async passwordRecoveryTest(@Body() body: PasswordRecoveryInputDto): Promise<null> {
-    const notification = await this.commandBus.execute<PasswordRecoveryCommand, ResultNotification<null>>(
-      new PasswordRecoveryCommand(body.email),
-    );
-    return notification.getData();
   }
 }
