+++ src/modules/auth/api/input-dto/password-recovery.input.dto.ts
@@ -15,6 +15,9 @@ export class PasswordRecoveryInputDto {
   })
   email: string;
 
+  /**
+   * Recaptcha token
+   */
   @IsString()
   @IsNotEmpty()
   @IsOptional()
