+++ src/modules/auth/api/view-dto/login-success.view.dto.ts
@@ -1,3 +1,6 @@
+/**
+ * Login success view DTO
+ */
 export type LoginSuccessViewDto = {
   accessToken: string;
 };
