+++ src/modules/auth/api/view-dto/password-recovery-view.dto.ts
@@ -1,3 +1,6 @@
+/**
+ * Password recovery view dto
+ */
 export class PasswordRecoveryViewDto {
   email: string;
 }
