+++ src/modules/auth/domain/password-recovery.entity.ts
@@ -9,15 +9,6 @@ export class PasswordRecoveryEntity {
 
   constructor() {}
 
-  static preparation(passRecovery: any): PasswordRecoveryEntity {
-    const passRecoveryEntity = new PasswordRecoveryEntity();
-    passRecoveryEntity.id = passRecovery.id;
-    passRecoveryEntity.email = passRecovery.email;
-    passRecoveryEntity.recoveryCode = passRecovery.recoveryCode;
-    passRecoveryEntity.expirationDate = passRecovery.expirationDate;
-    return passRecoveryEntity;
-  }
-
   static initCreate(email: string) {
     const passRecoveryEntity = new PasswordRecoveryEntity();
     passRecoveryEntity.email = email;
