+++ src/modules/posts/api/input-dto/create-post.input.dto.ts
@@ -2,18 +2,33 @@ import { ArrayMinSize, IsNotEmpty, IsString, MaxLength, ValidateNested } from 'c
 import { ApiProperty } from '@nestjs/swagger';
 import { Type } from 'class-transformer';
 
+/**
+ * @description Child metadata dto
+ */
 export class ChildMetadataDto {
+  /**
+   * @description Image upload id
+   */
   @IsString()
   @IsNotEmpty()
   uploadId: string;
 }
 
+/**
+ * @description Create post input dto
+ */
 export class CreatePostInputDto {
+  /**
+   * @description Post description (optional)
+   */
   @ApiProperty({ nullable: true, required: false })
   @MaxLength(500)
   @IsString()
   description: string;
 
+  /**
+   * @description Post children metadata
+   */
   @ValidateNested({ each: true })
   @Type(() => ChildMetadataDto)
   @ArrayMinSize(1)
