+++ src/modules/posts/api/input-dto/update-post.input.dto.ts
@@ -1,7 +1,13 @@
 import { IsString, MaxLength } from 'class-validator';
 import { ApiProperty } from '@nestjs/swagger';
 
+/**
+ * @description Update post input dto
+ */
 export class UpdatePostInputDto {
+  /**
+   * @description Post description (optional)
+   */
   @ApiProperty({ nullable: true, required: true })
   @MaxLength(500)
   @IsString()
