+++ src/modules/posts/api/posts-get.controller.ts
@@ -1,27 +1,28 @@
 import { Controller, Get, HttpCode, Param, ParseIntPipe, Query } from '@nestjs/common';
-import { CommandBus } from '@nestjs/cqrs';
 import { HTTP_Status } from '../../../main/enums/http-status.enum';
 import { ApiTags } from '@nestjs/swagger';
 import { IPostsQueryRepository } from '../infrastructure/posts-query.repository';
 import { PaginationPostsInputDto } from './input-dto/pagination-posts.input.dto';
-import { SwaggerDecoratorsByGetPosts } from '../swagger.posts.decorators';
+import { SwaggerDecoratorsByGetPosts } from '../swagger/swagger.posts.decorators';
 import { ApiOkResponsePaginated } from '../../../main/shared/api-ok-response-paginated';
 import { PostViewModel } from './view-models/post-view.dto';
 import { Paginated } from '../../../main/shared/paginated';
 
-// @ApiBearerAuth()
 @ApiTags('Posts')
-// @UseGuards(JwtAuthGuard)
 @Controller('posts')
 export class PostsGetController {
-  constructor(private readonly commandBus: CommandBus, private readonly postsQueryRepository: IPostsQueryRepository) {}
+  constructor(private readonly postsQueryRepository: IPostsQueryRepository) {}
 
+  /**
+   * Get posts by user id
+   * @param userId
+   * @param paginationInputModel
+   */
   @SwaggerDecoratorsByGetPosts()
   @ApiOkResponsePaginated(PostViewModel)
   @Get('/:userId')
   @HttpCode(HTTP_Status.OK_200)
   async getPosts(
-    // @CurrentUserId() userId: number,
     @Param(`userId`, ParseIntPipe) userId: number,
     @Query() paginationInputModel: PaginationPostsInputDto,
   ): Promise<Paginated<PostViewModel[]>> {
